+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "dataiku-dss" extension will be documented in this file.
 
+## 1.3.0 - To be defined when publishing the extension
+
+- Added support for project libraries.
+- Added ability to rename or move plugin content.
+
 ## 1.2.0 - May 27, 2020
 
 - Added support for wiki.